Code is secured, fix small errors

+++ src/main/java/com/mycompany/notes/service/impl/NotesServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
 
 @Service
@@ -29,17 +30,37 @@ public class NotesServiceImpl implements NotesService {
 
     @Override
     public Long insertNotes(Notes note){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        note.setIdNote(null);
+        note.setUsername(auth.getName());
+        note.setCreateTime(new Date());
+        note.setUpdateTime(null);
         return notesRepository.save(note).getIdNote();
     }
 
     @Override
     public Notes updateNotes(Notes note){
-        return notesRepository.save(note);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Notes note_ = notesRepository.findByUsernameAndIdNote(auth.getName(), note.getIdNote());
+        //check if the user is the owner
+        if(note_ != null) {
+            note.setUsername(auth.getName());
+            note.setCreateTime(note_.getCreateTime());
+            note.setUpdateTime(new Date());
+            return notesRepository.save(note);
+        }
+
+        return null;
     }
 
     @Override
     public void deleteNotes(Long idNote){
-        notesRepository.deleteById(idNote);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Notes note_ = notesRepository.findByUsernameAndIdNote(auth.getName(), idNote);
+        //check if the user is the owner
+        if(note_ != null) {
+            notesRepository.deleteById(idNote);
+        }
     }
 
 }
